move the logic out of the ArrowConverter helper function, create a non-static class to hold the state required when creating an ArrowSchema object | add 'add_metadata', turned off when outputting to Polars
