+++ src/common/arrow/arrow_wrapper.cpp
@@ -71,7 +71,7 @@ int ResultArrowArrayStreamWrapper::MyStreamGetSchema(struct ArrowArrayStream *st
 	auto my_stream = reinterpret_cast<ResultArrowArrayStreamWrapper *>(stream->private_data);
 	if (!my_stream->column_types.empty()) {
 		try {
-			ArrowConverter::ToArrowSchema(out, my_stream->column_types, my_stream->column_names,
+			ArrowConverter::ToArrowSchema(*out, my_stream->column_types, my_stream->column_names,
 			                              my_stream->result->client_properties);
 		} catch (std::runtime_error &e) {
 			my_stream->last_error = ErrorData(e);
@@ -97,7 +97,7 @@ int ResultArrowArrayStreamWrapper::MyStreamGetSchema(struct ArrowArrayStream *st
 		my_stream->column_names = result.names;
 	}
 	try {
-		ArrowConverter::ToArrowSchema(out, my_stream->column_types, my_stream->column_names,
+		ArrowConverter::ToArrowSchema(*out, my_stream->column_types, my_stream->column_names,
 		                              my_stream->result->client_properties);
 	} catch (std::runtime_error &e) {
 		my_stream->last_error = ErrorData(e);
