+++ tools/pythonpkg/src/arrow/arrow_export_utils.cpp
@@ -25,7 +25,7 @@ py::object ToArrowTable(const vector<LogicalType> &types, const vector<string> &
 	auto from_batches_func = pyarrow_lib_module.attr("Table").attr("from_batches");
 	auto schema_import_func = pyarrow_lib_module.attr("Schema").attr("_import_from_c");
 	ArrowSchema schema;
-	ArrowConverter::ToArrowSchema(&schema, types, names, options);
+	ArrowConverter::ToArrowSchema(schema, types, names, options);
 	auto schema_obj = schema_import_func(reinterpret_cast<uint64_t>(&schema));
 
 	return py::cast<duckdb::pyarrow::Table>(from_batches_func(batches, schema_obj));
