+++ tools/pythonpkg/src/pyresult.cpp
@@ -350,7 +350,8 @@ bool DuckDBPyResult::FetchArrowChunk(ChunkScanState &scan_state, py::list &batch
 	if (to_polars) {
 		QueryResult::DeduplicateColumns(names);
 	}
-	ArrowConverter::ToArrowSchema(&arrow_schema, query_result.types, names, query_result.client_properties);
+	ArrowConverter::ToArrowSchema(arrow_schema, query_result.types, names, query_result.client_properties,
+	                              to_polars == false);
 	TransformDuckToArrowChunk(arrow_schema, data, batches);
 	return true;
 }
